test: - move tests to EditableTask.test.ts
